refactor section parsing to be extensible for new doc types - create ParserDefinition class: base class for all doc type parsers. defines all section types. moved general stuff from Sections class here. - create ParserFactory class: determines what parser to use for a document - create DoDParser class, moved relevant logic from Sections class here - update tests to work w/ changes
also, initial upload of NavyParser
