+++ common/document_parser/lib/section_parse/__init__.py
@@ -1,3 +1,3 @@
 from .utils import *
-from .docx_parser import DocxParser
+from .parsers import *
 from .add_sections import add_sections
