+++ common/document_parser/lib/section_parse/parsers/dod_parser/__init__.py
@@ -1,0 +1,7 @@
+from .utils import *
+from .dod_parser import DoDParser
+
+# TODO here
+# TODO: tests for NavyParser class
+# TODO: tests for ParserDefinition class
+# TODO: tests for ParserFactory
