--- common/document_parser/lib/section_parse/tests/unit/__init__.py
@@ -1,3 +1,0 @@
-from .test_utils import UtilsTest
-from .test_sections import SectionsTest
-from .test_docx_parser import DocxParserTest
