+++ common/document_parser/lib/section_parse/utils/__init__.py
@@ -1,37 +1,2 @@
-from .utils import (
-    is_toc,
-    match_enclosure_num,
-    match_roman_numerals,
-    get_subsection,
-    next_section_num,
-    is_list_child,
-    is_space,
-    is_first_line_indented,
-    is_sentence_continuation,
-    is_alpha_list_item,
-    ends_with_continued,
-    ends_with_colon,
-    is_next_num_list_item,
-    match_num_list_item,
-    match_num_dot,
-    match_num_parentheses,
-    is_glossary_continuation,
-    starts_with_part,
-    match_section_num,
-    remove_strikethrough_text,
-    is_bold,
-    starts_with_bullet,
-    is_attachment_start,
-    starts_with_glossary,
-    match_attachment_num,
-    is_subsection_start_for_section_1,
-    get_subsection_of_section_1,
-)
-from .section_types import (
-    is_enclosure_continuation,
-    should_skip,
-    is_known_section_start,
-    is_child,
-    is_same_section_num,
-)
-from .sections import Sections
+from .docx_parser import DocxParser
+from .utils import is_alpha_list_item, match_num_list_item
